@@ -19,20 +19,20 @@
     "axios": "^0.27.2",
     "buefy": "^0.9.21",
     "codemirror": "^5.65.6",
-    "vue": "^2.6.14"
+    "vue": "^2.7.1"
   },
   "devDependencies": {
     "@types/codemirror": "^5.60.5",
-    "@types/node": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.30.0",
-    "@typescript-eslint/parser": "^5.30.0",
-    "eslint": "^8.18.0",
+    "@types/node": "^18.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "@vitejs/plugin-vue2": "^1.1.2",
+    "eslint": "^8.19.0",
     "eslint-plugin-vue": "^9.1.1",
     "sass": "^1.53.0",
     "typescript": "~4.7.4",
     "vite": "^2.9.13",
     "vite-plugin-compression": "^0.5.1",
-    "vite-plugin-vue2": "^2.0.1",
     "vue-axios": "^3.4.1"
   },
   "prettier": {
