@@ -26,7 +26,7 @@
 </template>
 
 <script lang="ts">
-import Vue, { PropType } from 'vue'
+import { defineComponent, PropType } from 'vue'
 
 import { FrameworkInfo } from '../api'
 
@@ -40,7 +40,7 @@ interface Data {
   count: number
 }
 
-export default Vue.extend({
+export default defineComponent({
   components: { CodeMirror, SaveButton },
 
   props: {
