@@ -1,7 +1,7 @@
 import { defineConfig } from 'vite'
 
 import compress from 'vite-plugin-compression'
-import { createVuePlugin as vue } from 'vite-plugin-vue2'
+import vue from '@vitejs/plugin-vue2'
 
 // https://vitejs.dev/config/
 export default defineConfig({
